+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.5.5-SNAPSHOT]
 
+- SortedFilteredList can now be assigned as items to TableView without calling `bindTo`
 - `DefaultErrorHandler.filter` listens to uncaught errors and can consume them to avoid the default error dialog.
 - `json.add(key, JsonModel)` automatically converts to JSON